% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfun.R
\name{dfun}
\alias{dfun}
\title{function to extract the overdispersion parameter from a quasi model}
\usage{
dfun(object)
}
\arguments{
\item{object}{a fitted model object}
}
\value{
dispersion parameter estimate
}
\description{
function to extract the overdispersion parameter from a quasi model
}
