% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fixed.R
\name{get_fixed}
\alias{get_fixed}
\title{make prediction plots for lm and glm objects using ggformula}
\usage{
get_fixed(data)
}
\arguments{
\item{data}{a data frame, or a fitted model object (from which the data will be extracted)}
}
\value{
a data frame with one row, containing the same variables as data, in which each variable takes on its median value (continuous variables) or its most commonly observed value (categorical variables)
}
\description{
make prediction plots for lm and glm objects using ggformula
}
