% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overdisp_fun.R
\name{overdisp_fun}
\alias{overdisp_fun}
\title{Estimate overdispersion for count data models}
\usage{
overdisp_fun(model)
}
\arguments{
\item{model}{a fitted model object}
}
\value{
The estimated overdispersion factor for the model.
}
\description{
From notes by Ben Bolker at https://bbolker.github.io/mixedmodels-misc/glmmFAQ.html#testing-for-overdispersioncomputing-overdispersion-factor
}
