% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para_boot.R
\name{para_boot}
\alias{para_boot}
\title{carry out parametric bootstrap to get CIs for model predictions}
\usage{
para_boot(model, new_data, predictor, nboot, conf_level = 0.95)
}
\arguments{
\item{model}{a fitted model object created by lm() or glm() or geeglm()}

\item{new_data}{data frame with data for which to make predictions and generate bootstrap CIs}

\item{predictor}{the covariate for which to make predictions. other predictors in the model will be held constant at their median value, or the most commonly observed value in the dataset.}

\item{nboot}{number of bootstrap iterations. Defaults to 1000. Ignored if boot is FALSE.}

\item{conf_level}{confidence level as a proportion, default is 0.95 for 95 percent confidence}
}
\value{
a data.frame with variables CIlower and CIupper, the lower and upper bounds of the parametric bootstrap CI
}
\description{
carry out parametric bootstrap to get CIs for model predictions
}
