% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_plot.R
\name{pred_plot}
\alias{pred_plot}
\title{make prediction plots for lm and glm objects using ggformula.}
\usage{
pred_plot(
  model,
  predictor,
  data = NULL,
  xlab = NULL,
  ylab = NULL,
  conf_level = 0.95,
  conf_int = TRUE,
  boot = FALSE,
  nboot = 1000,
  new_data_out = FALSE,
  ...
)
}
\arguments{
\item{model}{a fitted model object created by lm() or glm() or geeglm() or glmmTMB() or model.avg()}

\item{predictor}{the covariate for which to make predictions. other predictors in the model will be held constant at their median value, or the most commonly observed value in the dataset.}

\item{data}{The dataset to which the model was fitted. Only required (and only used) if the model input is an "averaging" object from model.avg().}

\item{xlab}{X axis label for plot (defaults to name of predictor variable)}

\item{ylab}{Y axis label for plot (defaults to "Predictions from Fitted Model")}

\item{conf_level}{confidence level as a proportion, default is 0.95 for 95 percent confidence}

\item{conf_int}{logical: should confidence intervals be shown (as error bars or confidence band)?}

\item{boot}{logical: should CIs be derived via a parametric bootstrap? Defaults to FALSE except if the model is a GEE, then defaults to TRUE.}

\item{nboot}{number of bootstrap iterations. Defaults to 1000. Ignored if boot is FALSE.}

\item{new_data_out}{logical: should data set used for predictions be output? If TRUE, result is a list including the plot object and the data frame.}

\item{...}{Additional arguments to be passed to plotting function}
}
\value{
A ggplot2 plot created using ggformul::gf_line() or gf_point()
}
\description{
make prediction plots for lm and glm objects using ggformula.
}
