% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x.quasipoisson.R
\name{x.quasipoisson}
\alias{x.quasipoisson}
\title{helper function for qdredge, not normally used directly by users. Modify a glm() output object so that it contains a quasipoisson fit but the AIC (likelihood) from the equivalent regular Poisson model}
\usage{
x.quasipoisson(...)
}
\arguments{
\item{...}{Additional arguments}
}
\value{
A fitted model object
}
\description{
helper function for qdredge, not normally used directly by users. Modify a glm() output object so that it contains a quasipoisson fit but the AIC (likelihood) from the equivalent regular Poisson model
}
