% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gee_anova.R
\name{gee_anova}
\alias{gee_anova}
\title{ANOVA for GEEs}
\usage{
gee_anova(model1)
}
\arguments{
\item{model1}{a fitted model object created by geepack::geeglm()}
}
\value{
a list containing the best fitted model (with terms with p-values above 0.05 removed via backward selection) and a table of p-values for all variables in model1
}
\description{
Carry out ANOVA and do backward model selection for GEEs fitted with geepack::geeglm()
}
