% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myQAIC.R
\name{myQAIC}
\alias{myQAIC}
\title{QAIC (easier) for quasi-poisson models}
\usage{
myQAIC(model, ...)
}
\arguments{
\item{model}{a fitted model object (or optionally more than one) with family=quasipoisson}

\item{\dots}{additional arguments to pass to MuMIn::QAIC()}
}
\value{
The QAIC for the model(s). See help for dredge() in package MuMIn for details.
}
\description{
QAIC (easier) for quasi-poisson models
}
