% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_plot.R
\name{pred_plot}
\alias{pred_plot}
\title{make prediction plots for lm and glm objects using ggformula.}
\usage{
pred_plot(model, predictor, xlab = NULL, ylab = NULL, conf_int = TRUE,
  ...)
}
\arguments{
\item{model}{a fitted model object created by lm() or glm()}

\item{predictor}{the covariate for which to make predictions. other predictors in the model will be held constant at their median value, or the most commonly observed value in the dataset.}

\item{xlab}{X axis label for plot (defaults to name of predictor variable)}

\item{ylab}{Y axis label for plot (defaults to "Predictions from Fitted Model")}

\item{conf_int}{logical: should 95 percent confidence intervals be shown (as error bars or confidence band)?}

\item{...}{Additional arguments to be passed to plotting function}

\item{new_data_out}{logical: should data set used for predictions be output? If TRUE, result is a list including the plot object and the data frame.}
}
\value{
A ggplot2 plot created using ggformula gf_line or gf_point
}
\description{
make prediction plots for lm and glm objects using ggformula.
}
